Continued fixing multi_color + movie num input
multi_color now accepts movie keyword arguments. The figure generated
now also prints the min and max in the title and omits the x and y
labels to save space. multi_color now also accepts a range argment
as the input. This way you can cycle through several times.

We have also updated the enter movie number prompt to offer numbers
instead of strings, e.g. [0, 1, 2] not ['000', '001', '002'].
This is more inline with how the movie num should be input.

This adresses Issues: #4 #5 and #6