Fixes dtype casting for tags
numpy 1.16 broke my dtype casting in dump.py
previously you could put a 6 field structure into the first 6 (3 in real, 3 velocity)
slots of a 7 (+ tags) field structure using numpy.concatenate, however now you get a
structure size match error.

To fix this we just make an empty array with the 7 field dtype and populated it.

I am not sure if this will lead to any increase in processing time but to quote
Donald Knuth: "premature optimization is the root of all evil"